backend/pyproject.toml
[project]
name = "sentry"
version = "0.1.0"
description = "Web API for Sentry, a crash-detection IoT device attachable to helmets to help riders inform loved ones in times of emergency"
authors = [
    {name = "Vulprit Prooze",email = "raileyalin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0,<7.0)",
    "django-ninja (>=1.5.1,<2.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.8,<0.15.0)"
]

# ===== Ruff Linter =========
[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ERA001"]

[tool.ruff.lint.per-file-ignores]
"*/migrations/*" = ["ALL"]
"*/settings/*" = ["ANN", "F401"]
"*/urls.py" = ["ANN"]
"*/wsgi.py" = ["ANN", "F401"]  # WSGI files
"*/asgi.py" = ["ANN", "F401"]  # ASGI files
"*/manage.py" = ["ANN", "F401", "PLC0415", "T201", "INP001"]
"*/__init__.py" = ["ALL"]


# ===== Pyright Configuration =====
[tool.pyright]
include = ["sentry"]
exclude = [
    "**/__pycache__",
    "**/migrations",
    "venv",
    "*/manage.py",
]
executionEnvironments = [
  { root = "sentry", pythonVersion = "3.12" }
]



# # === Dependency Management ===
# poetry add package                    # Add to production
# poetry add --group dev package        # Add to dev group
# poetry add --group test package       # Add to test group
# poetry remove package                 # Remove from production
# poetry remove --group dev package     # Remove from dev group
# poetry update                         # Update all
# poetry update package                 # Update specific package

# # === Installation ===
# poetry install                        # Install all
# poetry install --no-dev               # Production only
# poetry install --with dev             # With dev group

# # === Environment ===
# poetry shell                          # Activate virtual env
# poetry run <command>                  # Run command in env
# poetry env info                       # Show env info
# poetry env remove python              # Remove virtual env

# # === Information ===
# poetry show                           # List dependencies
# poetry show --group dev               # List dev group
# poetry show --outdated                # Show outdated
# poetry check                          # Validate pyproject.toml

# poetry env info                       # Check current environment


# ===== Ruff Commands ====
# Commands:
#   check    Run Ruff on the given files or directories
#   rule     Explain a rule (or all rules)
#   config   List or describe the available configuration options
#   linter   List all supported upstream linters
#   clean    Clear any caches in the current directory and any subdirectories
#   format   Run the Ruff formatter on the given files or directories
#   server   Run the language server
#   analyze  Run analysis over Python source code
#   version  Display Ruff's version
#   help     Print this message or the help of the given subcommand(s)

# Options:
#   -h, --help     Print help
#   -V, --version  Print version

# Log levels:
#   -v, --verbose  Enable verbose logging
#   -q, --quiet    Print diagnostics, but nothing else
#   -s, --silent   Disable all logging (but still exit with status code "1" upon detecting diagnostics)

# Global options:
#       --config <CONFIG_OPTION>  Either a path to a TOML configuration file (`pyproject.toml` or `ruff.toml`), or a TOML
#                                 `<KEY> = <VALUE>` pair (such as you might find in a `ruff.toml` configuration file)
#                                 overriding a specific configuration option. Overrides of individual settings using this
#                                 option always take precedence over all configuration files, including configuration
#                                 files that were also specified using `--config`
#       --isolated                Ignore all configuration files